Remove projects and add shortcutTeams to db
Remove the projects column as we no longer filter stories based on
project. Add a shortcutTeam column instead that is an array of teams
from shortcut IDs. I've made the distinction in code between `team` that
are teams in keeper, and `shortcutTeams` that are teams that come from
shortcut.

When creating a new story for an empty goal, we automatically create it
for the first team in the list, as most of the time you will only have
  one team selected anyway when using keeper. The list of projects when
  clicking "create story" is now populated by the available stories'
  projects. Unfortunately I couldn't find a relationship in the Shortcut
  API between teams and projects so this was the easiest way to list
  available projects.